app/requirement.txt
tortoise-orm
paho-mqtt
pydantic
asyncpg
gunicorn
uvicorn
fastapi
